Allow paying the joining fee after KYC approval
This deprecates refundEligible and removes the need to set and unset it
thus making the payJoiningFee function cheaper and denying KYC no longer
needs an onchain transaction, it would simply not call the signUpMember
function.

This commit also removes the onlyOwner modifier and the functions that
were using it. All of these functions were no longer used on mainnet and
neither they were tested or used in the script that deploys the whole
system from zero.